Create a `search-results` component that is now re-used on trending
