+++ app/components/search-results/group.js
