+++ app/templates/components/application/site-header.hbs
@@ -13,8 +13,9 @@
       </div>
 
       <ul class="nav navbar-nav pull-xs-right navbar-buttons">
+        {{! Search }}
         <li class="nav-item">
-          {{application/site-search class="site-search"}}
+          {{application/nav-search class="site-search"}}
         </li>
 
         {{#if session.hasUser}}
