--- app/templates/components/application/site-search.hbs
@@ -1,57 +1,0 @@
-{{one-way-input
-  update=(perform search)
-  id="search"
-  class=inputClass
-  placeholder="Search..."
-  autocomplete="off"
-}}
-<label for="search" class="site-search__icon" {{action (action (mut isOpened) true)}}>
-  <span class="icon">
-    {{svg-jar "search"}}
-  </span>
-</label>
-
-{{#if (and isOpened (not (is-empty query)))}}
-  {{! Wrap in div due to https://github.com/yapplabs/ember-wormhole/issues/66#issuecomment-263575168 }}
-  <div>
-    {{#epic-tooltip
-      id="search-popover"
-      target=(concat "#" elementId)
-      isStatic=true
-      singleInstance=true
-      attachment="top right"
-      targetAttachment="bottom right"
-      offset="-30px 0px"
-    }}
-      <div class="search--drop">
-        <p class="search--header">Search Results</p>
-        <div style="overflow: auto">
-          {{! anime group }}
-          {{application/site-search/group
-            group="anime"
-            items=groups.anime
-            isLoading=animeTask.isRunning
-            moreLink=(href-to "anime.index" (query-params text=query))
-            close=(action "close")
-          }}
-          {{! manga group }}
-          {{application/site-search/group
-            group="manga"
-            items=groups.manga
-            isLoading=mangaTask.isRunning
-            moreLink=(href-to "manga.index" (query-params text=query))
-            close=(action "close")
-          }}
-          {{! users group }}
-          {{application/site-search/group
-            group="users"
-            items=groups.users
-            isLoading=usersTask.isRunning
-            update=(action "updatePage" "users")
-            close=(action "close")
-          }}
-        </div>
-      </div>
-    {{/epic-tooltip}}
-  </div>
-{{/if}}
